Revert "Merge branch 'holiday_bug' of https://github.com/steve-the-bayesian/BOOM"
This reverts commit a949e173f7e073804d622e66e2bea51a6c235537, reversing
changes made to 1424b22c49df38e06ba585bc39a9a0e9c55a0f30.

+++ Models/Glm/BinomialLogitModel.cpp
@@ -1,4 +1,3 @@
-// Copyright 2018 Google LLC. All Rights Reserved.
 /*
   Copyright (C) 2005-2010 Steven L. Scott
 
@@ -16,37 +15,45 @@
   License along with this library; if not, write to the Free Software
   Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA
 */
-#include "Models/Glm/BinomialLogitModel.hpp"
-#include "cpputil/math_utils.hpp"
-#include "cpputil/report_error.hpp"
-#include "distributions.hpp"
-#include "stats/logit.hpp"
+#include <Models/Glm/BinomialLogitModel.hpp>
+#include <distributions.hpp>
+#include <stats/logit.hpp>
+#include <cpputil/math_utils.hpp>
+#include <cpputil/report_error.hpp>
 
-namespace BOOM {
+namespace BOOM{
   namespace {
     typedef BinomialLogitModel BLM;
     typedef BinomialRegressionData BRD;
   }  // namespace
 
   BLM::BinomialLogitModel(uint beta_dim, bool all)
-      : ParamPolicy(new GlmCoefs(beta_dim, all)), log_alpha_(0) {}
+      : ParamPolicy(new GlmCoefs(beta_dim, all)),
+        log_alpha_(0)
+  {}
 
   BLM::BinomialLogitModel(const Vector &beta)
-      : ParamPolicy(new GlmCoefs(beta)), log_alpha_(0) {}
+      : ParamPolicy(new GlmCoefs(beta)),
+        log_alpha_(0)
+  {}
 
-  BLM::BinomialLogitModel(const Ptr<GlmCoefs> &beta)
-      : ParamPolicy(beta), log_alpha_(0) {}
+  BLM::BinomialLogitModel(const Ptr<GlmCoefs> & beta)
+      : ParamPolicy(beta),
+        log_alpha_(0)
+  {}
 
   BLM::BinomialLogitModel(const Matrix &X, const Vector &y, const Vector &n)
-      : ParamPolicy(new GlmCoefs(X.ncol())), log_alpha_(0) {
-    int nr = nrow(X);
-    for (int i = 0; i < nr; ++i) {
-      uint yi = lround(y[i]);
-      uint ni = lround(n[i]);
-      NEW(BinomialRegressionData, dp)(yi, ni, X.row(i));
-      add_data(dp);
-    }
-  }
+      : ParamPolicy(new GlmCoefs(X.ncol())),
+        log_alpha_(0)
+      {
+        int nr = nrow(X);
+        for(int i = 0; i < nr; ++i){
+          uint yi = lround(y[i]);
+          uint ni = lround(n[i]);
+          NEW(BinomialRegressionData, dp)(yi, ni, X.row(i));
+          add_data(dp);
+        }
+      }
 
   BLM::BinomialLogitModel(const BLM &rhs)
       : Model(rhs),
@@ -58,9 +65,8 @@ namespace BOOM {
         PriorPolicy(rhs),
         log_alpha_(rhs.log_alpha_) {}
 
-  BLM *BinomialLogitModel::clone() const {
-    return new BinomialLogitModel(*this);
-  }
+  BLM* BinomialLogitModel::clone()const{
+    return new BinomialLogitModel(*this);}
 
   namespace {
     // Compute the probability of success (or failure) at a value of x.
@@ -70,12 +76,12 @@ namespace BOOM {
     //   success: If true, then the probability of success is
     //     returned.  Otherwise the probability of failure is returned.
     template <class VECTOR>
-    double logit_success_probability(const VECTOR &x, const GlmCoefs &beta,
-                                     bool success) {
+    double logit_success_probability(
+        const VECTOR &x, const GlmCoefs &beta, bool success) {
       double eta = beta.predict(x);
       return plogis(eta, 0, 1, success, false);
     }
-  }  // namespace
+  }
 
   double BLM::success_probability(const Vector &x) const {
     return logit_success_probability(x, coef(), true);
@@ -96,30 +102,29 @@ namespace BOOM {
     return logit_success_probability(x, coef(), false);
   }
 
-  double BLM::pdf(const Ptr<Data> &dp, bool logscale) const {
-    return pdf(DAT(dp), logscale);
-  }
+  double BLM::pdf(const Ptr<Data> & dp, bool logscale) const{
+    return pdf(DAT(dp), logscale);}
 
-  double BLM::pdf(const Data *dp, bool logscale) const {
+  double BLM::pdf(const Data * dp, bool logscale) const{
     const BinomialRegressionData *rd =
         dynamic_cast<const BinomialRegressionData *>(dp);
     return logp(rd->y(), rd->n(), rd->x(), logscale);
   }
 
-  double BLM::pdf(const Ptr<BRD> &dp, bool logscale) const {
+  double BLM::pdf(const Ptr<BRD> & dp, bool logscale) const{
     return logp(dp->y(), dp->n(), dp->x(), logscale);
   }
 
-  double BLM::logp_1(bool y, const Vector &x, bool logscale) const {
+  double BLM::logp_1(bool y, const Vector &x, bool logscale)const{
     double btx = predict(x);
     double ans = -lope(btx);
-    if (y) ans += btx;
+    if(y) ans += btx;
     return logscale ? ans : exp(ans);
   }
 
   // In many cases y and n will be set using integers, so they will
   // compare to integer literals exactly.  Only rarely are they non-integers.
-  double BLM::logp(double y, double n, const Vector &x, bool logscale) const {
+  double BLM::logp(double y, double n, const Vector &x, bool logscale)const{
     if (n == 0) {
       double ans = y == 0 ? 0 : negative_infinity();
       return logscale ? ans : exp(ans);
@@ -135,34 +140,34 @@ namespace BOOM {
     }
   }
 
-  double BLM::Loglike(const Vector &beta, Vector &g, Matrix &h, uint nd) const {
-    if (nd >= 2) return log_likelihood(beta, &g, &h);
-    if (nd == 1) return log_likelihood(beta, &g, 0);
+  double BLM::Loglike(const Vector &beta, Vector &g, Matrix &h, uint nd)const{
+    if(nd>=2) return log_likelihood(beta, &g, &h);
+    if(nd==1) return log_likelihood(beta, &g, 0);
     return log_likelihood(beta, 0, 0);
   }
 
-  double BLM::log_likelihood(const Vector &beta, Vector *g, Matrix *h,
-                             bool initialize_derivs) const {
+  double BLM::log_likelihood(const Vector & beta, Vector *g, Matrix *h,
+                             bool initialize_derivs)const{
     const BLM::DatasetType &data(dat());
     if (initialize_derivs) {
-      if (g) {
+      if (g){
         g->resize(beta.size());
-        *g = 0;
+        *g=0;
         if (h) {
           h->resize(beta.size(), beta.size());
-          *h = 0;
+          *h=0;
         }
       }
     }
     double ans = 0;
     bool all_coefficients_included = (xdim() == beta.size());
     const Selector &inc(coef().inc());
-    for (int i = 0; i < data.size(); ++i) {
+    for(int i = 0; i < data.size(); ++i){
       // y and n had been defined as uint's but y-n*p was computing
       // -n, which overflowed
-      double y = data[i]->y();
-      double n = data[i]->n();
-      const Vector &x(data[i]->x());
+      int y = data[i]->y();
+      int n = data[i]->n();
+      const Vector & x(data[i]->x());
       Vector reduced_x;
       if (!all_coefficients_included) {
         reduced_x = inc.select(x);
@@ -174,9 +179,9 @@ namespace BOOM {
       double loglike = dbinom(y, n, p, true);
       ans += loglike;
       if (g) {
-        g->axpy(X, y - n * p);  // g += (y-n*p) * x;
+        g->axpy(X, y-n*p);  // g += (y-n*p) * x;
         if (h) {
-          h->add_outer(X, X, -n * p * (1 - p));  // h += -npq * x x^T
+          h->add_outer(X, X, -n*p*(1-p)); // h += -npq * x x^T
         }
       }
     }
@@ -185,17 +190,19 @@ namespace BOOM {
 
   d2TargetFunPointerAdapter BLM::log_likelihood_tf() const {
     return d2TargetFunPointerAdapter(
-        [this](const Vector &x, Vector *g, Matrix *h, bool reset_derivs) {
-          return this->log_likelihood(x, g, h, reset_derivs);
-        });
+        [this](const Vector &x,
+               Vector *g,
+               Matrix *h,
+               bool reset_derivs) {
+          return this->log_likelihood(x, g, h, reset_derivs);});
   }
 
-  SpdMatrix BLM::xtx() const {
-    const std::vector<Ptr<BinomialRegressionData> > &d(dat());
+  SpdMatrix BLM::xtx()const{
+    const std::vector<Ptr<BinomialRegressionData> > & d(dat());
     uint n = d.size();
     uint p = d[0]->xdim();
     SpdMatrix ans(p);
-    for (uint i = 0; i < n; ++i) {
+    for(uint i=0; i<n; ++i){
       double n = d[i]->n();
       ans.add_outer(d[i]->x(), n, false);
     }
@@ -203,8 +210,8 @@ namespace BOOM {
     return ans;
   }
 
-  void BLM::set_nonevent_sampling_prob(double alpha) {
-    if (alpha <= 0 || alpha > 1) {
+  void BLM::set_nonevent_sampling_prob(double alpha){
+    if(alpha <=0 || alpha > 1){
       ostringstream err;
       err << "alpha (proportion of non-events retained in the data) "
           << "must be in (0,1]" << endl
@@ -214,6 +221,6 @@ namespace BOOM {
     log_alpha_ = std::log(alpha);
   }
 
-  double BLM::log_alpha() const { return log_alpha_; }
+  double BLM::log_alpha()const{return log_alpha_;}
 
 }  // namespace BOOM
